// src/audio-narrator/raindrop.gen.ts
// Code generated by raindrop build, DO NOT EDIT.
import { Ai, Annotation, App, Bucket, KvCache, Logger, MRNObject, ServiceStub, SqlDatabase, Tracer,  } from '@liquidmetal-ai/raindrop-framework';

export interface Env {
  _raindrop: {
    app: App;
  };
  AI: Ai;
  annotation: Annotation<Omit<MRNObject, 'type' | 'applicationName' | 'versionId'>>;
  API_GATEWAY: ServiceStub<import('../api-gateway/index.js').default>;
  AUDIO_STORAGE: Bucket;
  GOOGLE_CLOUD_CREDENTIALS: string;
  logger: Logger;
  mem: KvCache;
  STORY_GENERATOR: ServiceStub<import('../story-generator/index.js').default>;
  STORYTELLER_DB: SqlDatabase;
  tracer: Tracer;
}